chore: create the ui folder structure
Create the `/src/ui` folder to contain all components.

Create the `/src/ui/layout` folder to contain structural components.

Create the `/src/ui/shared` folder to contain reusable components.

Create the `/src/ui/views` folder to contain section components.

Add the `.gitkeep` file in all the created empty folders to include them in source control.

It is important to note that the name ui was chosen deliberately in place of components because the name is shorter.